move tests to a separate folder and remove canvas polyfill
