+++ .gitignore
@@ -25,4 +25,4 @@ website/i18n/*
 yarn-error.log
 local.log
 
-coverage/*
+tests/coverage/*
