+++ .prettierignore
@@ -1,2 +1,2 @@
 /dist
-/coverage
+/tests/coverage
