+++ package.json
@@ -9,7 +9,7 @@
     "start": "webpack serve --config webpack.config.js --mode development",
     "format": "prettier --write \"**/*.js\"",
     "postinstall": "patch-package",
-    "test": "jest"
+    "test": "jest --config tests/jest.config.js"
   },
   "files": [
     "dist"
@@ -33,7 +33,6 @@
     "@tensorflow/tfjs-node": "^4.17.0",
     "all-contributors-cli": "^6.26.1",
     "babel-jest": "^29.7.0",
-    "canvas": "^2.11.2",
     "cross-fetch": "^4.0.0",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.7.0",
@@ -60,7 +59,11 @@
     "@tensorflow/tfjs": "^4.2.0",
     "@tensorflow/tfjs-vis": "^1.5.1",
     "axios": "^1.3.4",
-    "canvas": "^2.11.2",
     "webpack-merge": "^5.9.0"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
   }
 }
