+++ tests/testingUtils/index.js
@@ -1,6 +1,3 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { createImageData, ImageData } from "canvas";
-
 export const asyncLoadImage = async (src) => {
   const img = new Image();
   if (src.startsWith("http")) {
@@ -26,11 +23,6 @@ export const randomImageData = (width = 200, height = 100) => {
     Math.floor(Math.random() * 256)
   );
   // Initialize a new ImageData object
-  return createImageData(array, width, height);
+  return { width, height, data: array };
 };
 
-export const polyfillImageData = () => {
-  if (!global.ImageData) {
-    global.ImageData = ImageData;
-  }
-};
