+++ tests/testingUtils/setupTests.js
@@ -1,4 +1,3 @@
-const { ImageData } = require("canvas");
 import '@tensorflow/tfjs-node'; // loads the tensorflow/node backend to the registry
 import crossFetch from 'cross-fetch';
 import * as tf from '@tensorflow/tfjs';
@@ -10,11 +9,6 @@ async function setupTests() {
   await tf.setBackend('tensorflow');
   tf.env().set('IS_BROWSER', false);
   tf.env().set('IS_NODE', true);
-
-  // Use the node-canvas ImageData polyfill
-  if (!global.ImageData) {
-    global.ImageData = ImageData;
-  }
 
   // Use cross-fetch as a polyfill for the browser fetch
   if (!global.fetch) {
